Devolve
Devolve keeps a record of all the transactions placed. It is a peer-to-peer payment & lending cryptocurrency that eliminates third parties. For example, if you needed a ride somewhere then you would typically call a transportation service, not anymore. With devolve, you can pay someone directly in the network to avoid fees. You can also do this for depositing money, insurance, for friends, whatever you want!